requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
openai==0.27.6
redis==4.5.4
tiktoken==0.4.0